Adding script to create the virtual environment
Ticket: None
Reviewed By: None
Testing Done: Created and tested a virtual environment

In order to simplify things in the pydotsim, a virtual environment
can be created with all the libraries setup to run pydotsim.
The default directory that the virtual environment is created in is
the .venv directory in the current working directory.  If you
call 'create_virtual_env.py' with the '-h' option, it will print
out help.