Adding new locking mechanism for UDP ports
Ticket: None
Reviewed By: None
Testing Done: Ran test simulation

Implement the mechanism for locking UDPs that are to be used in a
simulation.  If the directory with the file locks for the UDP ports
doesn't exist, one is created and the UDP ports from a given range
is scanned to see if they are free.  If they are free, a file
representing the port will be added.

When a simulation is created, the system will try and lock the UDP
port file via an 'ioctl'.  It will check to see if the file is
empty.  If it is, it will add its unique key into it and then
unlock the file.  Otherwise, it checks another port file.